// models/sales_cutting_sheet.model.ts
export type SalesCuttingSheetState = {
  data: {
    sals_cutt_id: string;
    sals_cutt_date: string;
    sals_cutt_company_name: string;
    sals_cutt_tel: string;
    sals_cutt_email: string;
    sals_cutt_doc_type: string;
    sals_cutt_printing_type: string;
    sals_cutt_amount: string;
    sals_cutt_quotation_request: string;
    sals_cutt_finished_size: string;
    sals_cutt_page: string;
    sals_cutt_text_paper: string;
    sals_cutt_printing: string;
    sals_cutt_text_coating: string;
    sals_cutt_printing_volume: string;
  };
};